DiscordBot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.28.0
aiohttp>=3.8.5
pydantic>=2.0.0